スタンプ編集時にも、正方形でないときはエラーが出るように
