+++ src/components/Settings/StampTab/StampItem.vue
@@ -48,6 +48,7 @@ import type { Stamp } from '@traptitech/traq'
 import useStateDiff from '../composables/useStateDiff'
 import { isValidStampName } from '/@/lib/validate'
 import { useToastStore } from '/@/store/ui/toast'
+import { imageSize } from './imageSize'
 
 type StampEditState = Pick<Stamp, 'name' | 'creatorId'>
 
@@ -80,32 +81,49 @@ const useStampEdit = (
 
   const editStamp = async () => {
     try {
-      isEditing.value = true
-      const promises = []
-      if (isStateChanged.value) {
-        promises.push(
-          apis.editStamp(props.stamp.id, {
-            name: diffKeys.value.includes('name') ? state.name : undefined,
-            creatorId: diffKeys.value.includes('creatorId')
-              ? state.creatorId
-              : undefined
-          })
-        )
+      if (!newImageData.value) {
+        addErrorToast('画像を選択してください')
+        return
       }
-      if (newImageData.value !== undefined) {
-        promises.push(apis.changeStampImage(props.stamp.id, newImageData.value))
+      const size = await imageSize(newImageData.value)
+      if (size.height !== size.width) {
+        addErrorToast('画像が正方形ではありません。編集してください')
+        return
       }
-      await Promise.all(promises)
-      afterSuccess()
+      try {
+        isEditing.value = true
+        const promises = []
+        if (isStateChanged.value) {
+          promises.push(
+            apis.editStamp(props.stamp.id, {
+              name: diffKeys.value.includes('name') ? state.name : undefined,
+              creatorId: diffKeys.value.includes('creatorId')
+                ? state.creatorId
+                : undefined
+            })
+          )
+        }
+        if (newImageData.value !== undefined) {
+          promises.push(
+            apis.changeStampImage(props.stamp.id, newImageData.value)
+          )
+        }
+        await Promise.all(promises)
+        afterSuccess()
 
-      addSuccessToast('スタンプを更新しました')
+        addSuccessToast('スタンプを更新しました')
+      } catch (e) {
+        // eslint-disable-next-line no-console
+        console.error('スタンプの編集に失敗しました', e)
+
+        addErrorToast(formatResizeError(e, 'スタンプの編集に失敗しました'))
+      }
+      isEditing.value = false
     } catch (e) {
       // eslint-disable-next-line no-console
-      console.error('スタンプの編集に失敗しました', e)
-
-      addErrorToast(formatResizeError(e, 'スタンプの編集に失敗しました'))
+      console.error('スタンプの作成に失敗しました', e)
+      addErrorToast(formatResizeError(e, 'スタンプの作成に失敗しました'))
     }
-    isEditing.value = false
   }
 
   return { isEditing, editStamp }
